+++ scripts/beatmap.js
@@ -253,7 +253,6 @@ async function analyseMap() {
             getSPSLowest() > (map.audio.duration < 120 ? 3.2 : 4.2) &&
             getSPSTotalPercDrop() < 60
         ) {
-            console.log(map.analysis.sps);
             arrText.push(
                 printHTMLBold(
                     `Minimum SPS not met (<${map.audio.duration < 120 ? '3.2' : '4.2'})`,
@@ -468,6 +467,7 @@ async function analyseDifficulty(charName, diff) {
     if (flag.tool.hbStaircase) {
         arrText.push(printHTMLBold('Hitbox staircase []', detectHitboxStaircase(diff._data, mapSettings)));
     }
+    arrText.push(printHTMLBold('Hitbox reverse staircase []', detectReverseStaircase(diff._data, mapSettings)));
     arrText.push(printHTMLBold('Improper window snap []', detectImproperWindowSnap(diff._data, mapSettings)));
     if (flag.tool.slowSlider) {
         arrText.push(