+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Changed
 
 -   Improved swing detection
+-   Cleaned up note function
 
 ## v1.5.7 [2021-05-01]